requirements.txt
beautifulsoup4==4.12.3
Flask==3.1.0
python-dotenv==1.0.1
selenium==4.27.1
gunicorn==23.0.0
psycopg2-binary==2.9.6
flask-sqlalchemy==2.5.1